+++ CaveStoryClone/CaveStoryClone/AnimatedTile.cpp
@@ -36,7 +36,7 @@ void AnimatedTile::Update(int elapsedTime)
 
 void AnimatedTile::Draw(Graphics &graphics)
 {
-	SDL_Rect destRect = { this->_position.x, this->_position.y, this->_size.x * globals::SPRITE_SCALE,  this->_size.y * globals::SPRITE_SCALE };
+	SDL_Rect destRect = { this->_position.x, this->_position.y, this->_size.x * (int)globals::SPRITE_SCALE,  this->_size.y * (int)globals::SPRITE_SCALE };
 	SDL_Rect sourceRect = { this->_tilesetPositions.at(this->_tileToDraw).x, this->_tilesetPositions.at(this->_tileToDraw).y, this->_size.x, this->_size.y };
 	graphics.BlitSurface(this->_tileset, &sourceRect, &destRect);
 }
